# LEAF-4121: Fulltext search