--- a/LEAF_Request_Portal/js/formSearch.js
+++ b/LEAF_Request_Portal/js/formSearch.js
@@ -473,6 +473,15 @@ var LeafFormSearch = function (containerID) {
                         }
                     });
                     empSel.initialize();
+                    let previousSelectedEmp = $("#" + prefixID + "widgetMat_" + widgetID).val();
+                    if(previousSelectedEmp != '') {
+                        if(type == 'empUID') {
+                            empSel.forceSearch(`#${previousSelectedEmp}`);
+                        }
+                        else {
+                            empSel.forceSearch(previousSelectedEmp);
+                        }
+                    }
                 },
             });
         } else {
@@ -505,6 +514,15 @@ var LeafFormSearch = function (containerID) {
                 }
             });
             empSel.initialize();
+            let previousSelectedEmp = $("#" + prefixID + "widgetMat_" + widgetID).val();
+            if(previousSelectedEmp != '') {
+                if(type == 'empUID') {
+                    empSel.forceSearch(`#${previousSelectedEmp}`);
+                }
+                else {
+                    empSel.forceSearch(previousSelectedEmp);
+                }
+            }
         }
     }
 
@@ -901,7 +919,7 @@ var LeafFormSearch = function (containerID) {
                         widgetID +
                         '" style="width: 140px" class="chosen" aria-label="categoryID">\
 	            		<option value="=">IS</option>\
-	            		<option value="!=">IS NOT</option>\
+	            		<option value="!=" selected>IS NOT</option>\
 	            	</select>'
                 );
                 url =
@@ -925,7 +943,7 @@ var LeafFormSearch = function (containerID) {
                         categories +=
                             '<option value="deleted">Cancelled</option>';
                         categories +=
-                            '<option value="resolved">Resolved</option>';
+                            '<option value="resolved" selected>Resolved</option>';
                         categories +=
                             '<option value="actionable">Actionable by me</option>';
                         //categories += '<option value="destruction">Scheduled for Destruction</option>';
@@ -1397,6 +1415,28 @@ var LeafFormSearch = function (containerID) {
                     },
                 });
                 break;
+            case "recordID":
+                $("#" + prefixID + "widgetCondition_" + widgetID).html(
+                    '<select id="' +
+                        prefixID +
+                        "widgetCod_" +
+                        widgetID +
+                        '" class="chosen" aria-label="condition" style="width: 55px">\
+		            		<option value="=">=</option>\
+		            		<option value=">">></option>\
+		            		<option value=">=">>=</option>\
+		            		<option value="<"><</option>\
+		            		<option value="<="><=</option>\
+		            	</select>'
+                );
+                $("#" + prefixID + "widgetMatch_" + widgetID).html(
+                    '<input type="text" aria-label="text" id="' +
+                        prefixID +
+                        "widgetMat_" +
+                        widgetID +
+                        '" style="width: 200px" />'
+                );
+                break;
             default:
                 $("#" + prefixID + "widgetCondition_" + widgetID).html(
                     '<select id="' +
@@ -1458,14 +1498,15 @@ var LeafFormSearch = function (containerID) {
             "widgetTerm_" +
             widgetCounter +
             '" style="width: 150px" class="chosen" aria-label="condition">\
-            				<option value="title">Title</option>\
-            				<option value="serviceID">Service</option>\
-            				<option value="dateSubmitted">Date Submitted</option>\
-            				<option value="categoryID">Type</option>\
-            				<option value="userID">Initiator</option>\
-            				<option value="dependencyID">Requirement</option>\
-            				<option value="stepID">Current Status</option>\
+                            <option value="stepID">Current Status</option>\
             				<option value="data">Data Field</option>\
+            				<option value="dateSubmitted">Date Submitted</option>\
+                            <option value="userID">Initiator</option>\
+            				<option value="serviceID">Service</option>\
+            				<option value="title">Title</option>\
+            				<option value="categoryID">Type</option>\
+                            <option value="recordID">Record ID</option>\
+            				<option value="dependencyID">Requirement</option>\
             				</select></td>\
 			            <td id="' +
             prefixID +