This follows [PR#2127](https://github.com/department-of-veterans-affairs/LEAF/pull/2127) and implements fulltext search for data fields within the portal.data table, which provides a better overall search experience. Backward compatibility in ./api/form/query is not impacted because the change only adds new search terms: MATCH ALL, MATCH, and NOT MATCH.

Using the original implementation as a baseline: Although the worst case performance is significantly better, this comes at a cost in best case performance and minor change in initial sort order. Average case performance also seems to be better when using the fulltext index.

| Condition | Before (s) | After (s) |
| :-- | :---: | :---: |
| Worst case performance (obscure search term "sterile pants") | 22.363 | 0.007 |
| Best case performance (common word "clinic" in dataset) | 0.008 | 0.122 |
| Average case performance (popular search term "FY23") | 0.341 | 0.004 |

Other UX changes:
- Changed default "Advanced Options" search view, and dropdown order
- Resolved issue where a selected employee appears blank after refreshing the page
- Resolved issue where a query using a NOT LIKE term ignored null fields

### Opportunities
- Explore whether OPTIMIZE table has any benefit
- Explore [innodb_ft_max_token_size](https://dev.mysql.com/doc/refman/8.0/en/innodb-parameters.html#sysvar_innodb_ft_max_token_size)
- Explore hybrid search UX using both LIKE and MATCH terms

### Potential Impact
- form/query and areas involving search
- The change in sort order could potentially cause confusion. This only impacts the homepage when an advanced search is applied, where the initial 50 results is sorted by relevance instead of the original date sort. Additionally this can be mitigated by simply clicking on "Show more records". Coordinate with the training team to prepare explanations for users.

### Testing
- [Experimental automated test added](https://github.com/mgaoVA/LEAF_test_experiments/commit/ac85fe53c3a1274d80a41748c94a7450a18397cd#diff-b7b1922b928b6c4ca9f0342a10188972bc7aa07897032f05faff1cb2b96f5234R56)
- Preprod: On a large site with at least 100,000 records, run a search for "Any standard data field CONTAINS FY23". The results should generally be the same as before.

### Deployment Prerequisites
- [ ] LEAF Infrastructure PR#109 must be merged and deployed
- [ ] Confirm that cache control is working properly for JS files